requirements.txt
kafka-python==2.0.2
requests==2.31.0
beautifulsoup4==4.12.2
flask==2.3.3
